app/requirements.txt
streamlit==1.46.0
snowflake-snowpark-python==1.40.0